requirements.txt
discord.py==2.1.0
python-dotenv==0.21.1
cogwatch==3.2.0
shopee-wrapper
tokopedia-wrapper
asyncpraw==7.6.1